Refactor currency validation workflow into modular nodes
Split the monolithic currency_validation.py into modular components: state management, data loading, phase 1 and phase 2 analysis, and label mapping. Introduced a Pydantic-based CurrencyAnalysisState for type safety and serialization, and moved development state saving to a dedicated utility. This improves maintainability, testability, and clarity of the workflow.
